pyproject.toml
[tool.poetry]
name = "fue-human-centered-design"
version = "0.1.0"
description = ""
authors = ["Alexander Schulz <schulz.alexander2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
openai = "^1.58.1"
python-decouple = "^3.8"
groq = "^0.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
